Update RemoteQueryException
-Rename RemoteQueryException to RemoteQueryHandlingException
-Add missing javadoc

#848
