+++ axon-server-connector/src/main/java/org/axonframework/axonserver/connector/query/RemoteQueryHandlingException.java
@@ -1,0 +1,57 @@
+package org.axonframework.axonserver.connector.query;
+
+import io.axoniq.axonserver.grpc.ErrorMessage;
+import org.axonframework.messaging.RemoteExceptionDescription;
+import org.axonframework.messaging.RemoteHandlingException;
+
+/**
+ * An AxonServer Exception which is thrown on a Query Handling exception.
+ *
+ * @author Marc Gathier
+ * @since 4.0
+ */
+public class RemoteQueryHandlingException extends RemoteHandlingException {
+
+    private final String errorCode;
+    private final String server;
+
+    /**
+     * Initialize a Query Handling exception from a remote source.
+     *
+     * @param errorCode a {@link String} defining the error code of this exception
+     * @param message   an {@link ErrorMessage} describing the exception
+     */
+    public RemoteQueryHandlingException(String errorCode, ErrorMessage message) {
+        super(new RemoteExceptionDescription(message.getDetailsList()));
+        this.errorCode = errorCode;
+        this.server = message.getLocation();
+    }
+
+    /**
+     * Return a {@link String} defining the error code.
+     *
+     * @return a {@link String} defining the error code
+     */
+    public String getErrorCode() {
+        return errorCode;
+    }
+
+
+    /**
+     * Return a {@link String} defining the location where the error originated.
+     *
+     * @return a {@link String} defining the location where the error originated
+     */
+    public String getServer() {
+        return server;
+    }
+
+    @Override
+    public String toString() {
+        return "RemoteQueryHandlingException{" +
+                "message=" + getMessage() +
+                ", errorCode='" + errorCode + '\'' +
+                ", location='" + server + '\'' +
+                '}';
+    }
+}
